food_stories/requirements.txt
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==2.2.8
django-filter==2.4.0
django-jet==1.0.8
django-modeltranslation==0.17.3
djangorestframework==3.12.4
idna==2.10
Markdown==3.3.4
oauthlib==3.1.1
Pillow==8.2.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
urllib3==1.26.6
